requirements.txt
flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
openai==1.6.1
pypdf==3.15.4
langchain==0.0.314
langchain-openai==0.0.2
faiss-cpu==1.7.4
python-socketio==5.9.0
uuid==1.30
gunicorn==20.1.0
numpy==1.24.2
# Optional RAG components - install manually if needed
chromadb==0.4.15
sentence-transformers==2.2.2
httpx==0.23.3
requests==2.28.1 